requirements.txt
# required python packages
kymatio
matplotlib
numpy
pytorch-ignite
ray[tune]
scikit-learn
tensorboardX
torch>=1.13.0+cu117
torchmetrics
torchvision